third_party/WebKit/LayoutTests/fast/css/invalidation/no-invalidation-set-local-style-expected.txt

Class selectors in rightmost compound selector should cause LocalStyleChange even if we cannot use invalidation sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(elm, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(elm, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(elm, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
